+++ tests/integration/annotation_import/test_model_run.py
@@ -14,8 +14,11 @@ def test_model_run(client, configured_project_with_label, rand_gen):
     model = client.create_model(data["name"], data["ontology_id"])
 
     name = rand_gen(str)
-    model_run = model.create_model_run(name)
+    config = {"batch_size": 100, "reruns": None}
+    model_run = model.create_model_run(name, config)
     assert model_run.name == name
+    assert model_run.training_metadata["batchSize"] == config["batch_size"]
+    assert model_run.training_metadata["reruns"] == config["reruns"]
     assert model_run.model_id == model.uid
     assert model_run.created_by_id == client.get_user().uid
 
@@ -30,6 +33,12 @@ def test_model_run(client, configured_project_with_label, rand_gen):
 
     fetch_model_run = client.get_model_run(model_run.uid)
     assert fetch_model_run == model_run
+
+
+def test_model_run_no_config(rand_gen, model):
+    name = rand_gen(str)
+    model_run = model.create_model_run(name)
+    assert model_run.name == name
 
 
 def test_model_run_delete(client, model_run):